// src/app/model/company.ts
export class Company {
    signStartTime: string;
    distance: number;
    addTime: Date;
    companyDesc: string;
    companyName: string;
    latitude: number;
    companyStatus: number;
    updateTime: Date;
    companyId: number;
    companyAddress: string;
    signEndTime: string;
    longitude: number;
}